# docs/openapi/be-plan-phase4.openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AgriDrone.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/farms": {
      "get": {
        "tags": [
          "Farm"
        ],
        "summary": "Lấy danh sách farm của tenant hiện tại.",
        "description": "Trả danh sách farm phân trang trong tenant từ access token. Chỉ Tenant\r\nAdmin hoặc Owner được phép xem catalog farm qua endpoint này.",
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Farm"
        ],
        "summary": "Tạo farm trong tenant hiện tại.",
        "description": "Tenant Admin hoặc Owner tạo farm với mã duy nhất trong tenant, thông\r\ntin vị trí và boundary GeoJSON sử dụng hệ tọa độ WGS84 (SRID 4326).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "410": {
            "description": "Legacy flow disabled (LegacyFlow.Disabled)."
          }
        },
        "deprecated": true
      }
    },
    "/api/farms/archived": {
      "get": {
        "tags": [
          "Farm"
        ],
        "summary": "Lấy danh sách farm đã archive của tenant hiện tại.",
        "description": "Chỉ Tenant Owner được phép xem danh sách. Response bao gồm thời điểm\r\narchive và version hiện tại để phục vụ thao tác restore an toàn.",
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}/archived": {
      "get": {
        "tags": [
          "Farm"
        ],
        "summary": "Lấy chi tiết một farm đã archive.",
        "description": "Chỉ Tenant Owner được phép xem. Farm phải thuộc tenant hiện tại và\r\nđang ở trạng thái archived; nếu không, API trả về 404.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}": {
      "get": {
        "tags": [
          "Farm"
        ],
        "summary": "Lấy chi tiết farm.",
        "description": "Trả thông tin farm thuộc tenant hiện tại khi người dùng có quyền đọc\r\nfarm thông qua tenant ownership hoặc farm membership.",
        "parameters": [
          {
            "name": "FarmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Farm"
        ],
        "summary": "Cập nhật thông tin farm.",
        "description": "Tenant Admin hoặc Owner cập nhật tên, địa chỉ, vị trí, boundary và\r\ndiện tích farm. ExpectedVersion được dùng để ngăn ghi đè cập nhật đồng thời.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmDetailRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmDetailRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmDetailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}/zones": {
      "post": {
        "tags": [
          "Farm"
        ],
        "summary": "Tạo zone trong farm.",
        "description": "Tenant Owner hoặc Farm Manager được assign tạo zone với mã duy nhất\r\ntrong farm. Boundary GeoJSON, nếu có, phải là Polygon SRID 4326 hợp lệ.\r\nZone mới được kích hoạt và bắt đầu ở version 1.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Farm"
        ],
        "summary": "Lấy danh sách zone theo farm.",
        "description": "Trả các zone chưa archive thuộc farm và tenant hiện tại. Kết quả được\r\nlọc theo FarmAccessScope và ZoneAssignment đang hoạt động của người dùng.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}/zones/{zoneId}": {
      "get": {
        "tags": [
          "Farm"
        ],
        "summary": "Lấy chi tiết zone.",
        "description": "Trả zone thuộc đúng farm và tenant hiện tại khi người dùng có quyền\r\nqua FarmAccessScope hoặc ZoneAssignment. Response bao gồm version phục\r\nvụ optimistic concurrency cho các thao tác cập nhật sau này.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Farm"
        ],
        "summary": "Cập nhật thông tin một zone.",
        "description": "Owner hoặc Farm Manager có quyền trên đúng zone được phép cập nhật tên,\r\nboundary và diện tích. ExpectedVersion ngăn ghi đè thay đổi đồng thời.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}/zones/{zoneId}/archive": {
      "put": {
        "tags": [
          "Farm"
        ],
        "summary": "Archive một zone và giữ nguyên lịch sử liên quan.",
        "description": "Chỉ Tenant Owner được phép thực hiện. Yêu cầu bị từ chối nếu zone còn\r\nMission đang hoạt động hoặc Field Task chưa đóng.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveZoneRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveZoneRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveZoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}/archive": {
      "put": {
        "tags": [
          "Farm"
        ],
        "summary": "Archive một farm và giữ nguyên lịch sử liên quan.",
        "description": "Chỉ Tenant Owner được phép thực hiện. Farm chỉ được archive khi không\r\ncòn Zone, Mission hoặc Field Task đang hoạt động.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveFarmRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveFarmRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveFarmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/farms/{farmId}/restore": {
      "put": {
        "tags": [
          "Farm"
        ],
        "summary": "Restore một farm và giữ nguyên lịch sử liên quan.",
        "description": "Chỉ Tenant Owner được phép thực hiện. Farm chỉ được restore khi Farm đang ở trạng thái archived.",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreFarmRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreFarmRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreFarmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveFarmRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "ArchiveZoneRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "CreateFarmRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "boundary": {
            "$ref": "#/components/schemas/GeoJsonPolygonRequest"
          },
          "centerPoint": {
            "$ref": "#/components/schemas/GeoJsonPointRequest"
          },
          "areaHectares": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateZoneRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "boundary": {
            "$ref": "#/components/schemas/GeoJsonPolygonRequest"
          },
          "areaHectares": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GeoJsonPointRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GeoJsonPolygonRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "type": "number",
                  "format": "double"
                }
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RestoreFarmRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateFarmDetailRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "boundary": {
            "$ref": "#/components/schemas/GeoJsonPolygonRequest"
          },
          "centerPoint": {
            "$ref": "#/components/schemas/GeoJsonPointRequest"
          },
          "areaHectares": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateZoneRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "boundary": {
            "$ref": "#/components/schemas/GeoJsonPolygonRequest"
          },
          "areaHectares": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}